# blocpress-render/src/main/resources/META-INF/openapi.yml
openapi: 3.0.3
info:
  title: Blocpress Render API
  version: 1.0.0
  description: >
    REST API for document generation from ODT templates and JSON data.

    **Stateless rendering** (`POST /api/render/template`) generates a document from a
    template you supply inline — either as a multipart file upload or as a base64-encoded
    JSON request. No authentication required.

    **Name-based rendering** (`POST /api/render/{name}`) renders a document using an
    approved template stored in the production template store. The latest active version
    of the named template is used. Requires a valid JWT Bearer token.

    **Authentication:** The server ships with a built-in dev key (issuer `https://blocpress.dev`).
    To use your own identity provider, override the environment variables
    `MP_JWT_VERIFY_PUBLICKEY` and `MP_JWT_VERIFY_ISSUER` when starting the container.

servers:
  - url: http://localhost:8080

security:
  - BearerJWT: []

tags:
  - name: Rendering
    description: Document generation endpoints

paths:
  /api/render/template:
    post:
      tags: [ Rendering ]
      operationId: renderDocumentStateless
      summary: Generate document from ODT template (stateless)
      description: >
        Submit an ODT template and JSON data to generate a document.
        No authentication required — the template is provided inline with each request.

        Two request formats are supported, selected via the `Content-Type` header:

        - **`multipart/form-data`**: Upload the ODT file as `template` field, JSON data
          as `data` field. Use the `Accept` header to specify the desired output format.

        - **`application/json`**: Submit a base64-encoded ODT template with JSON data
          and output type in a single JSON body.
      security: []
      parameters:
        - name: Accept
          in: header
          required: false
          schema:
            type: string
            enum:
              - application/pdf
              - application/rtf
              - application/vnd.oasis.opendocument.text
            example: application/pdf
          description: >
            Desired output format for multipart requests. One of: `application/pdf`,
            `application/rtf`, `application/vnd.oasis.opendocument.text`.
            Ignored for JSON requests (use `outputType` field instead).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - template
                - data
              properties:
                template:
                  type: string
                  format: binary
                  description: ODT template file
                data:
                  type: string
                  description: JSON template data as string (key-value pairs matching template fields)
          application/json:
            schema:
              $ref: '#/components/schemas/RenderRequest'
      responses:
        "200":
          description: Generated document
          headers:
            Content-Disposition:
              description: Download filename
              schema:
                type: string
                example: 'attachment; filename="document.pdf"'
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            application/rtf:
              schema:
                type: string
                format: binary
            application/vnd.oasis.opendocument.text:
              schema:
                type: string
                format: binary
        "400":
          description: Bad request — missing template, invalid JSON, or invalid template structure
        "406":
          description: Unsupported output format
        "500":
          description: Internal server error

  /api/render/{name}:
    post:
      tags: [ Rendering ]
      operationId: renderDocumentByName
      summary: Generate document from stored template
      description: >
        Render a document using an approved template from the production template store.
        The latest active version of the named template is used. Only APPROVED templates
        are available for rendering. Template content is cached for performance.
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Template name (must exist and be in APPROVED status)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenderByNameRequest'
      responses:
        "200":
          description: Generated document
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            application/rtf:
              schema:
                type: string
                format: binary
            application/vnd.oasis.opendocument.text:
              schema:
                type: string
                format: binary
        "400":
          description: Bad request — invalid JSON data
        "401":
          description: Unauthorized — missing or invalid JWT
        "403":
          description: Template not approved for rendering
        "404":
          description: Template not found
        "500":
          description: Internal server error

components:
  schemas:
    RenderRequest:
      type: object
      required:
        - template
        - data
        - outputType
      properties:
        template:
          type: string
          format: byte
          description: Base64-encoded ODT template
        data:
          type: object
          description: Template data (key-value pairs matching template fields)
        outputType:
          type: string
          enum:
            - pdf
            - rtf
            - odt
          description: Desired output format
    RenderByNameRequest:
      type: object
      required:
        - data
        - outputType
      properties:
        data:
          type: object
          description: Template data (key-value pairs matching template fields)
        outputType:
          type: string
          enum:
            - pdf
            - rtf
            - odt
          description: Desired output format
  securitySchemes:
    BearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT